requirements.txt
flask
reportlab
PyPDF2
Pillow
numpy
matplotlib